reinforce + implement DisplayConfig
when block data got loaded, performs a check to restrict stored values with configured ones.

+++ src/main/java/me/srrapero720/waterframes/util/FrameTools.java
@@ -59,7 +59,7 @@ public class FrameTools {
         return is;
     }
 
-    public static double minFloat(float a, float b) {
+    public static float minFloat(float a, float b) {
         if (a != a)
             return a;   // a is NaN
         if ((a == 0.0f) &&
